upd: build config / img remote paths
