+++ app/data/banners/banners.json
@@ -20,25 +20,25 @@
     "caption" : "RetroStyle",
     "btn": "Жми на 1!",
     "img" :[
-        {"background": "http://localhost:8090/banner/img/background.jpg"},
-        {"blick_glasses":"http://localhost:8090/banner/img/blick_glasses.png"},
-        {"blick_earring":"http://localhost:8090/banner/img/blick_earring.png"},
-        {"blick_hair1":"http://localhost:8090/banner/img/blick_hair1.png"},
-        {"blick_jacket":"http://localhost:8090/banner/img/blick_jacket.png"},
-        {"blick_jacket2":"http://localhost:8090/banner/img/blick_jacket2.png"},
-        {"blick_summer":"http://localhost:8090/banner/img/blick_summer.png"},
-        {"car_ligth":"http://localhost:8090/banner/img/car_ligth.png"},
-        {"hair":"http://localhost:8090/banner/img/hair.png"},
-        {"hair2":"http://localhost:8090/banner/img/hair2.png"},
-        {"hair3":"http://localhost:8090/banner/img/hair3.png"},
-        {"jaket1":"http://localhost:8090/banner/img/jaket1.png"},
-        {"jaket2":"http://localhost:8090/banner/img/jaket2.png"},
-        {"mask_car":"http://localhost:8090/banner/img/mask_car.png"},
-        {"mask_floor":"http://localhost:8090/banner/img/mask_floor.png"},
-        {"open_O":"http://localhost:8090/banner/img/open_O.png"},
-        {"palm1":"http://localhost:8090/banner/img/palm1.png"},
-        {"palm2":"http://localhost:8090/banner/img/palm2.png"},
-        {"woman":"http://localhost:8090/banner/img/woman.png"}
+        {"background": "/public/banner/img/background.jpg"},
+        {"blick_glasses":"/public/banner/img/blick_glasses.png"},
+        {"blick_earring":"/public/banner/img/blick_earring.png"},
+        {"blick_hair1":"/public/banner/img/blick_hair1.png"},
+        {"blick_jacket":"/public/banner/img/blick_jacket.png"},
+        {"blick_jacket2":"/public/banner/img/blick_jacket2.png"},
+        {"blick_summer":"/public/banner/img/blick_summer.png"},
+        {"car_ligth":"/public/banner/img/car_ligth.png"},
+        {"hair":"/public/banner/img/hair.png"},
+        {"hair2":"/public/banner/img/hair2.png"},
+        {"hair3":"/public/banner/img/hair3.png"},
+        {"jaket1":"/public/banner/img/jaket1.png"},
+        {"jaket2":"/public/banner/img/jaket2.png"},
+        {"mask_car":"/public/banner/img/mask_car.png"},
+        {"mask_floor":"/public/banner/img/mask_floor.png"},
+        {"open_O":"/public/banner/img/open_O.png"},
+        {"palm1":"/public/banner/img/palm1.png"},
+        {"palm2":"/public/banner/img/palm2.png"},
+        {"woman":"/public/banner/img/woman.png"}
     ],
     "effect" : 12,
     "transitionType" : 3,
