--- app/style.css
@@ -1,30 +1,0 @@
-#c {
-    width: 1600px;
-    height: 547px;
-}
-
-#caption1{
-    position: absolute;
-    top: 250px;
-    left: 160px;
-    color: azure;
-    font-size: 48px;
-}
-
- #btn {
-    position: absolute;
-    top: 340px;
-    left: 160px;
-    
-    background-color: rgb(255, 238, 0);
-    color: rgb(0, 0, 0);
-    font-size: 24px;
-    padding: 18px 32px;
-    border: none;
-    cursor: pointer;
-    border-radius: 5px;
-  }
-  
-.btn:hover {
-    background-color: rgb(233, 194, 20);
-  }
