+++ index.html
@@ -2,8 +2,6 @@
 <html lang="en">
   <head>
     <meta charset="UTF-8" />
-    <link rel="icon" type="image/svg+xml" href="/vite.svg" />
-    <link rel = 'stylesheet' href = './app/style.css'>
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Vite App</title>
     <style>
@@ -13,6 +11,7 @@
       }
       #wrap {
         width: 1900px;
+        max-width: 100%;
         margin: 0 auto;
         display: flex;
         justify-content: center;
@@ -20,11 +19,11 @@
     </style>
   </head>
   <body>
-<!--    <canvas id='c'></canvas>-->
-<!--    <div id='caption1'>Тут будет подпись к баннеру</div>-->
-<!--    <button id="btn">Давай жми!</button>-->
+    <!--    <canvas id='c'></canvas>-->
+    <!--    <div id='caption1'>Тут будет подпись к баннеру</div>-->
+    <!--    <button id="btn">Давай жми!</button>-->
     <div id="wrap">
-      <div id="banner-c"/>
+      <div id="banner-c" />
     </div>
     <script type="module" src="/index.js"></script>
   </body>
