+++ package.json
@@ -25,6 +25,7 @@
   "author": "bb-team",
   "license": "ISC",
   "dependencies": {
+    "@rollup/plugin-replace": "^4.0.0",
     "@vitejs/plugin-legacy": "^2.0.1",
     "eslint-config-airbnb": "^19.0.4",
     "eslint-plugin-standard": "^5.0.0",
